+++ template/jsx/src/index.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+import * as styles from "./index.pcss";
+
+const Index: React.FC = (props) => {
+    return (
+        <div className={styles.mydiv}>
+            <div>div</div>
+            <span>span</span>
+        </div>
+    );
+};
+
+export { Index };